process: move deprecation warning initialization into pre_execution.js
Since this is only necessary when user code execution is expected.
