+++ lib/internal/bootstrap/node.js
@@ -229,29 +229,6 @@ Object.defineProperty(process, 'argv0', {
 });
 process.argv[0] = process.execPath;
 
-const { deprecate } = NativeModule.require('internal/util');
-{
-  // Install legacy getters on the `util` binding for typechecking.
-  // TODO(addaleax): Turn into a full runtime deprecation.
-  const pendingDeprecation = getOptionValue('--pending-deprecation');
-  const utilBinding = internalBinding('util');
-  const types = NativeModule.require('internal/util/types');
-  for (const name of [
-    'isArrayBuffer', 'isArrayBufferView', 'isAsyncFunction',
-    'isDataView', 'isDate', 'isExternal', 'isMap', 'isMapIterator',
-    'isNativeError', 'isPromise', 'isRegExp', 'isSet', 'isSetIterator',
-    'isTypedArray', 'isUint8Array', 'isAnyArrayBuffer'
-  ]) {
-    utilBinding[name] = pendingDeprecation ?
-      deprecate(types[name],
-                'Accessing native typechecking bindings of Node ' +
-                'directly is deprecated. ' +
-                `Please use \`util.types.${name}\` instead.`,
-                'DEP0103') :
-      types[name];
-  }
-}
-
 // process.allowedNodeEnvironmentFlags
 Object.defineProperty(process, 'allowedNodeEnvironmentFlags', {
   get() {
@@ -272,6 +249,8 @@ Object.defineProperty(process, 'allowedNodeEnvironmentFlags', {
   enumerable: true,
   configurable: true
 });
+
+const { deprecate } = NativeModule.require('internal/util');
 // process.assert
 process.assert = deprecate(
   perThreadSetup.assert,
