+++ lib/internal/bootstrap/pre_execution.js
@@ -2,6 +2,45 @@
 
 const { getOptionValue } = require('internal/options');
 
+// In general deprecations are intialized wherever the APIs are implemented,
+// this is used to deprecate APIs implemented in C++ where the deprecation
+// utitlities are not easily accessible.
+function initializeDeprecations() {
+  const { deprecate } = require('internal/util');
+  const pendingDeprecation = getOptionValue('--pending-deprecation');
+
+  // DEP0103: access to `process.binding('util').isX` type checkers
+  // TODO(addaleax): Turn into a full runtime deprecation.
+  const utilBinding = internalBinding('util');
+  const types = require('internal/util/types');
+  for (const name of [
+    'isArrayBuffer',
+    'isArrayBufferView',
+    'isAsyncFunction',
+    'isDataView',
+    'isDate',
+    'isExternal',
+    'isMap',
+    'isMapIterator',
+    'isNativeError',
+    'isPromise',
+    'isRegExp',
+    'isSet',
+    'isSetIterator',
+    'isTypedArray',
+    'isUint8Array',
+    'isAnyArrayBuffer'
+  ]) {
+    utilBinding[name] = pendingDeprecation ?
+      deprecate(types[name],
+                'Accessing native typechecking bindings of Node ' +
+                'directly is deprecated. ' +
+                `Please use \`util.types.${name}\` instead.`,
+                'DEP0103') :
+      types[name];
+  }
+}
+
 function initializeClusterIPC() {
   // If this is a worker in cluster mode, start up the communication
   // channel. This needs to be done before any user code gets executed
@@ -75,6 +114,7 @@ function loadPreloadModules() {
 }
 
 module.exports = {
+  initializeDeprecations,
   initializeClusterIPC,
   initializePolicy,
   initializeESMLoader,
