+++ lib/internal/main/check_syntax.js
@@ -4,6 +4,7 @@
 // instead of actually running the file.
 
 const {
+  initializeDeprecations,
   initializeClusterIPC,
   initializePolicy,
   initializeESMLoader,
@@ -21,6 +22,7 @@ const {
 } = require('internal/modules/cjs/helpers');
 
 // TODO(joyeecheung): not every one of these are necessary
+initializeDeprecations();
 initializeClusterIPC();
 initializePolicy();
 initializeESMLoader();
