+++ lib/internal/main/eval_stdin.js
@@ -3,6 +3,7 @@
 // Stdin is not a TTY, we will read it and execute it.
 
 const {
+  initializeDeprecations,
   initializeClusterIPC,
   initializePolicy,
   initializeESMLoader,
@@ -14,6 +15,7 @@ const {
   readStdin
 } = require('internal/process/execution');
 
+initializeDeprecations();
 initializeClusterIPC();
 initializePolicy();
 initializeESMLoader();
