+++ lib/internal/main/eval_string.js
@@ -4,6 +4,7 @@
 // `--interactive`.
 
 const {
+  initializeDeprecations,
   initializeClusterIPC,
   initializePolicy,
   initializeESMLoader,
@@ -13,6 +14,7 @@ const { evalScript } = require('internal/process/execution');
 const { addBuiltinLibsToObject } = require('internal/modules/cjs/helpers');
 
 const source = require('internal/options').getOptionValue('--eval');
+initializeDeprecations();
 initializeClusterIPC();
 initializePolicy();
 initializeESMLoader();
