+++ lib/internal/main/repl.js
@@ -4,6 +4,7 @@
 // the main module is not specified and stdin is a TTY.
 
 const {
+  initializeDeprecations,
   initializeClusterIPC,
   initializePolicy,
   initializeESMLoader,
@@ -14,6 +15,7 @@ const {
   evalScript
 } = require('internal/process/execution');
 
+initializeDeprecations();
 initializeClusterIPC();
 initializePolicy();
 initializeESMLoader();
