+++ lib/internal/main/run_main_module.js
@@ -1,12 +1,14 @@
 'use strict';
 
 const {
+  initializeDeprecations,
   initializeClusterIPC,
   initializePolicy,
   initializeESMLoader,
   loadPreloadModules
 } = require('internal/bootstrap/pre_execution');
 
+initializeDeprecations();
 initializeClusterIPC();
 initializePolicy();
 initializeESMLoader();
