+++ lib/internal/main/worker_thread.js
@@ -4,6 +4,7 @@
 // message port.
 
 const {
+  initializeDeprecations,
   initializeClusterIPC,
   initializeESMLoader,
   loadPreloadModules
@@ -55,6 +56,7 @@ port.on('message', (message) => {
     if (manifestSrc) {
       require('internal/process/policy').setup(manifestSrc, manifestURL);
     }
+    initializeDeprecations();
     initializeClusterIPC();
     initializeESMLoader();
     loadPreloadModules();
